many bug fixes, premuim added

+++ BotHandler.py
@@ -12,14 +12,14 @@ from os import remove, path
 
 from telethon.tl.types import DocumentAttributeVideo
 from telethon import TelegramClient
-from telegram import InputMediaVideo
+
 from threading import Thread, enumerate
 from telegram.ext import Updater, CommandHandler, MessageHandler, Filters, CallbackQueryHandler
 from asyncio import new_event_loop, set_event_loop
 
 
 
-home = path.dirname(path.realpath(__file__))
+home = path.dirname(path.realpath(__file__)) + "/"
 
 
 
@@ -56,6 +56,12 @@ def thread_help(update, context):
             update.message.reply_text('لطفا برای استفاده از ربات در کانال ما جوین شوید. :)', reply_markup=keyboards.ForceJoinKeyboard, reply_to_message_id=update.message.message_id)
             return
 
+def thread_premium(update, context):
+        update.message.reply_text("how to use the bot : \n\n ....", reply_markup=keyboards.buymenu)
+
+        update.message.reply_text(
+            "لطفا بعد از خرید از صفحه انجام تراکنش یا پیام برداشت از حساب عکس گرفته و برای ربات بفرستید تا حساب شما شارژ شود.",
+            reply_markup=keyboards.cancelbuy)
 
 
 def thread_link_manager(update, context):
@@ -64,80 +70,77 @@ def thread_link_manager(update, context):
         chat_id = update.message.chat_id
         link    = update.message.text
 
+
         if not methods.channel_checker(context, chat_id):
             update.message.reply_text('لطفا برای استفاده از ربات در کانال ما جوین شوید. :)', reply_markup=keyboards.ForceJoinKeyboard, reply_to_message_id=update.message.message_id)
             return
-        
+
+
+        if link == "❌ لغو خرید ❌":
+
+            update.message.reply_text(f"سلام {update.message.chat.first_name}, به ربات دانلودر توییتر خوش آومدی \nلینک توییتت رو بفرست اینجا تا برات فیلم هاشو بفرستم",
+                                       reply_markup=keyboards.RemoveKeys)
 
 
 
 
-
-        if (link.startswith("https://x.com") or link.startswith("https://twitter.com")):      # twitter section
+        elif (link.startswith("https://x.com") or link.startswith("https://twitter.com")):      # twitter section
 
             url, caption = methods.create_url(context, link)
-
             if not url:
                 update.message.reply_text("داخل این لینک هیچ فیلمی پیدا نکردم") 
                 return
-
             wait_message = context.bot.send_message(chat_id=chat_id, text="داره دانلود میشه صبر کن یکم ...", reply_to_message_id=update.message.message_id)
-
-
-
-            file_names = [f"{home}/cache/twitter/{chat_id}_{item}.mp4" for item in range(len(url))]
-  
+            file_names = [f"{home}cache/twitter/{chat_id}_{item}.mp4" for item in range(len(url))]
             i = 0
             for item in url:
                 methods.download_video(item, file_names[i])
                 i += 1
-
             context.bot.edit_message_text(chat_id=chat_id, message_id=wait_message.message_id, text="دانلود شد, الان میفرستمش برات")
-            
             if len(file_names) == 1: # if its one video in the tweet
-                context.bot.send_video(chat_id=chat_id, video=open(file_names[0], 'rb'), caption=caption, reply_markup=keyboards.SponsorKeyboard)
-
+                run_filesender(file_sender(chat_id, file_names[0], caption))
             else: # if its a group video
-                media_files = []
-                lastitem = file_names.pop()
-                media_files.append(InputMediaVideo(open(lastitem, "rb"), caption=caption))            
-
-                for item in file_names:
-                    media_files.append(InputMediaVideo(open(item, 'rb')))
-                context.bot.send_media_group(chat_id=chat_id, media=media_files)
-                
-                context.bot.send_message(chat_id=chat_id, text="""
-                                                                ( •_•)                       (•_• )
-                                                                ( ง )ง                       ୧( ୧ )
-                                                                /︶\                          /︶\\
-                                                                """, reply_markup=keyboards.SponsorKeyboard)
-
-
+                run_filesender(file_sender(chat_id, file_names, caption))
             context.bot.deleteMessage(chat_id=chat_id, message_id=wait_message.message_id)
             db.AddUsageNum(chat_id)
-
-
+            context.bot.send_message(chat_id=chat_id, text="""------ ( *︾▽︾)""", reply_markup=keyboards.SponsorKeyboard)
 
 
         elif link.startswith("https://www.youtube.com") or link.startswith("https://youtu.be") or link.startswith("https://youtube.com"):      # youtube section
-
             wait_message = context.bot.send_message(chat_id=chat_id, text="در حال پردازش ...", reply_to_message_id=update.message.message_id)
-
             data, working = methods.youtube_getinfo(link)
-
             if data and working:
                 context.bot.edit_message_text(chat_id=chat_id, message_id=wait_message.message_id, text=data["title"] + "\nlength : " + data["length"], reply_markup=keyboards.CreateKey(data))
             elif not working:
-
                 context.bot.edit_message_text(chat_id=chat_id, message_id=wait_message.message_id, text="this video cant be downloaded")
                 context.bot.send_message(chat_id=creds.Admin, text=data)
 
 
 
+
+        else: #dwonload from direct link
+            file_size = methods.get_file_size(link)
+
+            if file_size > 100 :
+                print(db.is_prem(chat_id))
+                if not db.is_prem(chat_id):        
+                    context.bot.send_message(chat_id=chat_id, text="حجم فایل مورد نظر بیشتر از ۱۰۰ مگابایت است")
+                    return
+                
+            wait_message = context.bot.send_message(chat_id=chat_id, text="در حال دانلود ...", reply_to_message_id=update.message.message_id)
+            file = methods.downloader(link)
+            context.bot.edit_message_text(chat_id=chat_id, message_id=wait_message.message_id, text="فایل دانلود شد,  در حال آپلود ...")
+            run_filesender(file_sender(chat_id, home + "cache/other/" + file))
+            context.bot.deleteMessage(chat_id=chat_id, message_id=wait_message.message_id)
+
+
+            context.bot.send_message(chat_id=chat_id, text="""------ ( *︾▽︾)""", reply_markup=keyboards.SponsorKeyboard)
+
     except Exception as error:
         
         context.bot.send_message(chat_id=chat_id, text="ی مشکلی پیش اومد ببشید, دوباره بفرست برام شاید تونستم")
         context.bot.send_message(chat_id=creds.Admin, text=f"error in link_manager by {chat_id}\nerror:\n{error}")
+        print(error)
 
 
     finally:
@@ -150,24 +153,32 @@ def thread_link_manager(update, context):
         #     pass
 
 
+async def file_sender(chat_id, file_path, caption=None):
 
 
-async def file_sender(chat_id, file_path, caption, duration):
-    try : 
-        async with TelegramClient("telethon", creds.ApiId, creds.ApiHash) as client:
-            attributes = [DocumentAttributeVideo(duration=duration, w=1280, h=720)]
-            await client.send_file(chat_id, file_path, attributes=attributes, caption=caption)    #TODO all formats files video audio ...
+    async with TelegramClient(home + "telethon", creds.ApiId, creds.ApiHash) as client:
+
+        try:
+            await client.send_message(chat_id, message=caption, file=file_path)
+
+        except Exception as error:
+            await client.send_message(chat_id, "ی مشکلی پیش اومد ببشید, دوباره بفرست برام شاید تونستم")
+            await client.send_message(creds.Admin, f"error in file_sender by {chat_id}\nerror:\n{error}")
+            print("--------")
+            print(error)
 
 
-    except Exception as error:
-        print("error in file sender", str(error))
-    finally:
+        finally:
+            if isinstance(file_path, list):
+                for file in file_path:
+                    if path.exists(file):
+                        remove(file)
+            else:
+                if path.exists(file_path):
+                    remove(file_path)
 
-        if path.isfile(file_path):
-            remove(file_path)
 
-
-def run_in_thread(coro):
+def run_filesender(coro):
     loop = new_event_loop()
     set_event_loop(loop)
     loop.run_until_complete(coro)
@@ -192,8 +203,8 @@ def thread_callback(update, context):
             if command == 'joined':
                 if methods.channel_checker(context, chat_id):
                     query.edit_message_text("ربات فعال شد الان میتونی استفاده کنی ")
-                    # thread_link_manager(query, context)
-                    # print(query.message.reply_to_message.text)
+                    thread_link_manager(query, context)
+
                 else:
                     if chat_id == creds.Admin :
                         query.answer("اوففف ادمینن")        
@@ -203,16 +214,25 @@ def thread_callback(update, context):
 
             elif command.split("-")[0] == "youtube" :   # youtube manager
 
+                generator = methods.youtube_getvideo(query.message.reply_to_message.text, command.split("-")[1])
+
+
+                if next(generator) > 100:
+
+                    context.bot.send_message(chat_id=chat_id, text="حجم فایل مورد نظر بیشتر از ۱۰۰ مگابایت است")
+                    return
+
                 query.edit_message_text("دانلود شروع شد ...")
-                file_path, title, duration = methods.youtube_getvideo(query.message.reply_to_message.text, command.split("-")[1])
+                file_path, title = next(generator)
     
                 query.edit_message_text("دانلود تمام شد , در حال آپلود...")
 
-                run_in_thread(file_sender(chat_id, file_path, title, duration))
+                run_filesender(file_sender(chat_id, file_path, title))
                 
                 context.bot.delete_message(chat_id=chat_id, message_id=query.message.message_id)
 
-
+                if path.isfile(file_path):
+                    remove(file_path)
 
             # if admin  #TODO bring admin down
             elif chat_id == creds.Admin :
@@ -224,13 +244,58 @@ def thread_callback(update, context):
                 elif command == "db":
                     context.bot.send_document(chat_id=chat_id, document=open(home+'db.sqlite', "rb"))
 
-                return
+
+
+
+
+                elif command.startswith("month"):
+
+                    month = int(command.split("-")[1])
+                    user = query.message.caption
+                    
+                    
+                    try:
+                        db.add_prem(user, month)
+                        if month == 0 :
+                            context.bot.send_message(chat_id=user, text=f"اعتبار حساب شما افزایش پیدا نکرد، از صحت عکس ارسالی مطمئن شوید.")
+                        else:
+                            context.bot.send_message(chat_id=user, text=f"حساب شا {month} ماه شارژ شد")
+                
+                        context.bot.edit_message_caption(
+                            caption=f"{query.data} added to {user}", 
+                            chat_id=chat_id, message_id=query.message.message_id)
+                        
+                    except Exception as error:
+                        context.bot.send_message(chat_id=chat_id, text=f"error in add month\n error : {error}")
+
 
         except Exception as error:     
+            print(error)
             context.bot.send_message(chat_id=chat_id, text="ی مشکلی پیش اومد ببشید, دوباره بفرست برام شاید تونستم")
-            context.bot.send_message(chat_id=creds.Admin, text=f"error in link_manager by {chat_id}\nerror:\n{error}")
+            context.bot.send_message(chat_id=creds.Admin, text=f"error in call back manager by {chat_id}\nerror:\n{error}")
            
-           
+
+
+def thread_forward(update, context):
+
+    chat_id = update.message.chat_id
+
+    try:
+        
+        photo_id = update.message.photo[-1].file_id
+        pic = context.bot.send_photo(chat_id=creds.Admin, photo=photo_id, caption=chat_id, reply_markup=keyboards.admin_addmenu)
+        if update.message.caption:
+            context.bot.send_message(chat_id=creds.Admin, text=update.message.caption,
+                                    reply_to_message_id=pic.message_id)
+        update.message.reply_text("در حال برسی, چند دقیقه صبر کنید ...", reply_to_message_id=update.message.message_id)
+        return
+
+    except Exception as error:
+        update.message.reply_text("مشکلی در سیستم پیش آمد, لطفا دوباره تلاش کنید.")
+        context.bot.send_message(chat_id=creds.Admin, text=f"error in thread forward by {update.message.caption}\nerror : {error}")
+
+
+
 #########################################################################################
 #########################################################################################
 
@@ -247,6 +312,12 @@ def help(update, context):
 def callback(update, context):
     Thread(target=thread_callback, args=(update, context, )).start()
 
+def premium(update, context):
+    Thread(target=thread_premium, args=(update, context, )).start()
+
+def forward_photo(update, context):
+    Thread(target=thread_forward, args=(update, context,)).start()
+
 def go_live():
     print("going live...")
     while True:
@@ -255,7 +326,9 @@ def go_live():
             updater.dispatcher.add_handler(CommandHandler('start', start))
             updater.dispatcher.add_handler(CommandHandler('restart', start))
             updater.dispatcher.add_handler(CommandHandler('help', help))
+            updater.dispatcher.add_handler(CommandHandler('premium', premium))
             updater.dispatcher.add_handler(CallbackQueryHandler(callback))
+            updater.dispatcher.add_handler(MessageHandler(Filters.photo, forward_photo))
             updater.dispatcher.add_handler(MessageHandler(Filters.text, link_manager))
             updater.start_polling()     
             print("bot is live.")
